feat: LocalWebRTCRoomの型と実装を新規作成し、LocalWebRTCHostSDKでの利用を更新
