+++ packages/local-webrtc-browser-sdk/src/sdk/local-web-rtc-room.ts
@@ -1,0 +1,33 @@
+/** ローカルWebRTC ルーム */
+export type LocalWebRTCRoom = {
+  /** ルームID */
+  readonly roomID: string;
+
+  /**
+   * マッチングするまで待機する
+   * @returns マッチングしたら発火するPromise
+   */
+  waitUntilMatching: () => Promise<void>;
+};
+
+/** ローカルWebRTCルームの実装 */
+export class LocalWebRTCRoomImpl implements LocalWebRTCRoom {
+  /** ルームID */
+  roomID: string;
+
+  /**
+   * コンストラクタ
+   * @param roomID ルームID
+   */
+  constructor(roomID: string) {
+    this.roomID = roomID;
+  }
+
+  /**
+   * マッチングするまで待機する
+   * @returns マッチングしたら発火するPromise
+   */
+  async waitUntilMatching(): Promise<void> {
+    // TODO マッチングの実装
+  }
+}
